Move JSGLR2Variants.Variant to JSGLR2Variant; move enum JSGLR2Variants to JSGLR2Variant.Preset
