+++ org.spoofax.jsglr2.integration/src/main/java/org/spoofax/jsglr2/integration/Sdf3ToParseTable.java
@@ -84,7 +84,7 @@ public class Sdf3ToParseTable {
         // TODO: use the parse table variant in the parse table generator
         return new ParseTable(normalizedGrammar, new ParseTableConfiguration(false, false, true, false, false));
     }
-    
+
     public IParseTable getLayoutSensitiveParseTable(ParseTableVariant variant, String sdf3Resource) throws Exception {
         NormGrammar normalizedGrammar = normalizedGrammarFromSDF3("grammars/" + sdf3Resource);
