+++ org.spoofax.jsglr2/src/test/java/org/spoofax/jsglr2/JSGLR2VariantTest.java
@@ -4,16 +4,16 @@ import org.junit.Test;
 import org.metaborg.parsetable.IParseTable;
 import org.spoofax.jsglr2.parser.ParserVariant;
 
-public class VariantTest {
+public class JSGLR2VariantTest {
 
     /**
      * This test will throw an IllegalStateException in JSGLR2Variants.getParser if the preset variants are invalid.
      * 
-     * @see JSGLR2Variants.Variant#getJSGLR2(IParseTable)
+     * @see JSGLR2Variant#getJSGLR2(IParseTable)
      * @see ParserVariant#getParser(IParseTable)
      */
     @Test public void testPresetVariants() {
-        for(JSGLR2Variants enumValue : JSGLR2Variants.values()) {
+        for(JSGLR2Variant.Preset enumValue : JSGLR2Variant.Preset.values()) {
             enumValue.variant.getJSGLR2(null);
         }
     }
